requirements.txt
langchain
langchain-core
langchain-groq
streamlit
python-dotenv